third_party/blink/web_tests/fast/lists/ol-reversed-dynamic-expected.txt

This tests that changing the value of an item updates markers accordingly.

10 Ten
9 Nine
8 Eight
7 Seven
6 Six


This tests that resetting the value of an item updates markers accordingly.

5 Five
4 Four
3 Three
2 Two
1 One


This tests that changing the start value of the list updates markers accordingly.

20 Twenty
19 Nineteen
18 Eighteen
17 Seventeen
16 Sixteen


This tests that removing the custom start value of the list updates markers accordingly.

5 Five
4 Four
3 Three
2 Two
1 One


This tests that changing the custom start value from 1 to "" works.

5 Five
4 Four
3 Three
2 Two
1 One
